Added check_response_status function. Post method returns data dictionary
