--- mailcamp/errors.py
@@ -1,7 +1,0 @@
-"""
-In this file the errors are being initialized
-"""
-
-
-class MailCampError(Exception):
-    pass
